Add PUT/users Validations

+++ api/src/controllers/user/deleteUserById.controller.js
@@ -1,9 +1,19 @@
 const { Users } = require("../../db")
+const { isValidUUID } = require("../../utils")
 
 const deleteUserById = async (req, res) => {
 
     try {
         const { id } = req.params
+
+        if (!isValidUUID(id))
+            return res.status(404).json({ error: "id must be UUID format" })
+
+        const userFound = await Users.findByPk(id)
+
+        if (!userFound)
+            return res.status(404).json({ error: "User not found" });
+
         await Users.destroy({
             where: {
                 userID: id
